Moved core.flows and core.contract tests
