+++ core-tests/src/test/resources/isolated.jar
