--- core/src/test/kotlin/net/corda/core/flows/FinalityFlowTests.kt
@@ -1,103 +1,0 @@
-package net.corda.core.flows
-
-import com.natpryce.hamkrest.and
-import com.natpryce.hamkrest.assertion.assertThat
-import net.corda.core.flows.mixins.WithFinality
-import net.corda.core.flows.mixins.WithFinality.FinalityInvoker
-import net.corda.core.identity.Party
-import net.corda.core.internal.cordapp.CordappResolver
-import net.corda.core.transactions.SignedTransaction
-import net.corda.core.transactions.TransactionBuilder
-import net.corda.core.utilities.getOrThrow
-import net.corda.finance.POUNDS
-import net.corda.finance.contracts.asset.Cash
-import net.corda.finance.issuedBy
-import net.corda.testing.core.*
-import net.corda.testing.internal.matchers.flow.willReturn
-import net.corda.testing.internal.matchers.flow.willThrow
-import net.corda.testing.node.internal.*
-import org.assertj.core.api.Assertions.assertThat
-import org.junit.After
-import org.junit.Test
-
-class FinalityFlowTests : WithFinality {
-    companion object {
-        private val CHARLIE = TestIdentity(CHARLIE_NAME, 90).party
-    }
-
-    override val mockNet = InternalMockNetwork(cordappsForAllNodes = listOf(FINANCE_CONTRACTS_CORDAPP, enclosedCordapp()))
-
-    private val aliceNode = makeNode(ALICE_NAME)
-
-    private val notary = mockNet.defaultNotaryIdentity
-
-    @After
-    fun tearDown() = mockNet.stopNodes()
-
-    @Test
-    fun `finalise a simple transaction`() {
-        val bob = createBob()
-        val stx = aliceNode.issuesCashTo(bob)
-
-        assertThat(
-            aliceNode.finalise(stx, bob.info.singleIdentity()),
-                willReturn(
-                        requiredSignatures(1)
-                                and visibleTo(bob)))
-    }
-
-    @Test
-    fun `reject a transaction with unknown parties`() {
-        // Charlie isn't part of this network, so node A won't recognise them
-        val stx = aliceNode.issuesCashTo(CHARLIE)
-
-        assertThat(
-            aliceNode.finalise(stx),
-                willThrow<IllegalArgumentException>())
-    }
-
-    @Test
-    fun `allow use of the old API if the CorDapp target version is 3`() {
-        val oldBob = createBob(cordapps = listOf(tokenOldCordapp()))
-        val stx = aliceNode.issuesCashTo(oldBob)
-        val resultFuture = CordappResolver.withCordapp(targetPlatformVersion = 3) {
-            @Suppress("DEPRECATION")
-            aliceNode.startFlowAndRunNetwork(FinalityFlow(stx)).resultFuture
-        }
-        resultFuture.getOrThrow()
-        assertThat(oldBob.services.validatedTransactions.getTransaction(stx.id)).isNotNull()
-    }
-
-    @Test
-    fun `broadcasting to both new and old participants`() {
-        val newCharlie = mockNet.createNode(InternalMockNodeParameters(legalName = CHARLIE_NAME))
-        val oldBob = createBob(cordapps = listOf(tokenOldCordapp()))
-        val stx = aliceNode.issuesCashTo(oldBob)
-        val resultFuture = aliceNode.startFlowAndRunNetwork(FinalityInvoker(
-                stx,
-                newRecipients = setOf(newCharlie.info.singleIdentity()),
-                oldRecipients = setOf(oldBob.info.singleIdentity())
-        )).resultFuture
-        resultFuture.getOrThrow()
-        assertThat(newCharlie.services.validatedTransactions.getTransaction(stx.id)).isNotNull()
-        assertThat(oldBob.services.validatedTransactions.getTransaction(stx.id)).isNotNull()
-    }
-
-    private fun createBob(cordapps: List<TestCordappInternal> = emptyList()): TestStartedNode {
-        return mockNet.createNode(InternalMockNodeParameters(legalName = BOB_NAME, additionalCordapps = cordapps))
-    }
-
-    private fun TestStartedNode.issuesCashTo(recipient: TestStartedNode): SignedTransaction {
-        return issuesCashTo(recipient.info.singleIdentity())
-    }
-
-    private fun TestStartedNode.issuesCashTo(other: Party): SignedTransaction {
-        val amount = 1000.POUNDS.issuedBy(info.singleIdentity().ref(0))
-        val builder = TransactionBuilder(notary)
-        Cash().generateIssue(builder, amount, other, notary)
-        return services.signInitialTransaction(builder)
-    }
-
-    /** "Old" CorDapp which will force its node to keep its FinalityHandler enabled */
-    private fun tokenOldCordapp() = cordappWithPackages("com.template").copy(targetPlatformVersion = 3)
-}
