+++ settings.gradle
@@ -15,6 +15,7 @@ include 'finance:contracts'
 include 'finance:workflows'
 include 'isolated'
 include 'core'
+include 'core-tests'
 include 'docs'
 include 'node-api'
 include 'node'
